Refactor EditableCourseAssignment2 fixed Assignment editor state clearing bug
